Move GroundFirestore into new schema package
